// src/shared/stock.generated.ts
export const STOCK: StockData = {
  families: {
// ...
    '--dsw-static-deepseek': [
      { name: '--dsw-static-deepseek-50', L: 0.962819, C: 0.016092, h: 262.753108 },
      { name: '--dsw-static-deepseek-100', L: 0.944071, C: 0.023683, h: 262.094743 },
      { name: '--dsw-static-deepseek-200', L: 0.910606, C: 0.042914, h: 263.690057 },
      { name: '--dsw-static-deepseek-300', L: 0.838567, C: 0.078161, h: 271.052078 },
      { name: '--dsw-static-deepseek-400', L: 0.70435, C: 0.152924, h: 260.841785 },
      { name: '--dsw-static-deepseek-450', L: 0.647252, C: 0.185831, h: 265.192933 },
      { name: '--dsw-static-deepseek-500', L: 0.589558, C: 0.179339, h: 263.017915 },
      { name: '--dsw-static-deepseek-600', L: 0.529123, C: 0.122491, h: 264.900799 },
      { name: '--dsw-static-deepseek-700-delete', L: 0.430638, C: 0.11603, h: 264.449821 },
      { name: '--dsw-static-deepseek-800', L: 0.375524, C: 0.048033, h: 263.734259 },
      { name: '--dsw-static-deepseek-900', L: 0.312461, C: 0.033131, h: 262.883436 },
    ],
    '--dsw-static-green': [
      { name: '--dsw-static-green-100', L: 0.96685, C: 0.027104, h: 157.383944 },
      { name: '--dsw-static-green-400', L: 0.768359, C: 0.165207, h: 152.341516 },
      { name: '--dsw-static-green-500', L: 0.722746, C: 0.192007, h: 149.57933 },
      { name: '--dsw-static-green-900', L: 0.330377, C: 0.042212, h: 155.272517 },
    ],
    '--dsw-static-neutral': [
      { name: '--dsw-static-neutral-00', L: 1, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-50', L: 0.985104, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-100', L: 0.970151, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-150', L: 0.946105, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-200', L: 0.921906, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-250', L: 0.89449, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-300', L: 0.869944, C: 0, h: 89.875564 },
      { name: '--dsw-static-neutral-400', L: 0.71768, C: 0.003693, h: 247.875298 },
      { name: '--dsw-static-neutral-500', L: 0.605694, C: 0.008346, h: 260.730455 },
      { name: '--dsw-static-neutral-550', L: 0.513371, C: 0.006752, h: 264.506231 },
      { name: '--dsw-static-neutral-600', L: 0.449304, C: 0.003486, h: 264.52618 },
      { name: '--dsw-static-neutral-700', L: 0.356542, C: 0.001703, h: 286.313052 },
      { name: '--dsw-static-neutral-800', L: 0.28094, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-850', L: 0.248568, C: 0.003737, h: 286.124712 },
      { name: '--dsw-static-neutral-900', L: 0.168416, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-1000', L: 0, C: 0, h: 0 },
    ],
    '--dsw-static-neutral-bluish': [
      { name: '--dsw-static-neutral-bluish-00', L: 1, C: 0, h: 89.875563 },
      { name: '--dsw-static-neutral-bluish-50', L: 0.984625, C: 0.001706, h: 247.838921 },
      { name: '--dsw-static-neutral-bluish-60', L: 0.972666, C: 0.001711, h: 247.839212 },
      { name: '--dsw-static-neutral-bluish-75', L: 0.963197, C: 0.003431, h: 247.858549 },
      { name: '--dsw-static-neutral-bluish-100', L: 0.947975, C: 0.006271, h: 255.47577 },
      { name: '--dsw-static-neutral-bluish-150', L: 0.942524, C: 0.008692, h: 264.520986 },
      { name: '--dsw-static-neutral-bluish-200', L: 0.92147, C: 0.012933, h: 266.695715 },
      { name: '--dsw-static-neutral-bluish-300', L: 0.864629, C: 0.006036, h: 239.832264 },
      { name: '--dsw-static-neutral-bluish-400', L: 0.761415, C: 0.010249, h: 252.834142 },
      { name: '--dsw-static-neutral-bluish-500', L: 0.693986, C: 0.014857, h: 258.356051 },
      { name: '--dsw-static-neutral-bluish-600', L: 0.61562, C: 0.011546, h: 261.777237 },
      { name: '--dsw-static-neutral-bluish-700', L: 0.50761, C: 0.010067, h: 248.026295 },
      { name: '--dsw-static-neutral-bluish-750', L: 0.39037, C: 0.008827, h: 268.429729 },
      { name: '--dsw-static-neutral-bluish-800', L: 0.33265, C: 0.003761, h: 264.513307 },
      { name: '--dsw-static-neutral-bluish-850', L: 0.293901, C: 0.003582, h: 286.177291 },
      { name: '--dsw-static-neutral-bluish-875', L: 0.256549, C: 0.00185, h: 286.26857 },
      { name: '--dsw-static-neutral-bluish-900', L: 0.22254, C: 0.001918, h: 286.242826 },
      { name: '--dsw-static-neutral-bluish-950', L: 0.196601, C: 0.003967, h: 286.029186 },
      { name: '--dsw-static-neutral-bluish-1000', L: 0.177275, C: 0.008873, h: 264.318285 },
    ],
    '--dsw-static-red': [
      { name: '--dsw-static-red-50', L: 0.970528, C: 0.012947, h: 17.379699 },
      { name: '--dsw-static-red-100', L: 0.935641, C: 0.030896, h: 17.717165 },
      { name: '--dsw-static-red-400', L: 0.668239, C: 0.187411, h: 23.667152 },
      { name: '--dsw-static-red-500', L: 0.636834, C: 0.207849, h: 25.331328 },
      { name: '--dsw-static-red-600', L: 0.59697, C: 0.238189, h: 28.623543 },
      { name: '--dsw-static-red-900', L: 0.297906, C: 0.106797, h: 26.670928 },
    ],
  },
  hex: {
// ...
    '--dsw-static-blue-500': '#3b82f6',
    '--dsw-static-blue-50p': '#eaf3ff',
    '--dsw-static-blue-600': '#2563eb',
    '--dsw-static-blue-75': '#e5f0ff',
    '--dsw-static-blue-800': '#1e40af',
    '--dsw-static-blue-900': '#0e3074',
    '--dsw-static-blue-950': '#172554',
    '--dsw-static-deepseek-100': '#e4edfd',
    '--dsw-static-deepseek-200': '#d3e2ff',
    '--dsw-static-deepseek-300': '#b7c8fe',
    '--dsw-static-deepseek-400': '#679efe',
    '--dsw-static-deepseek-450': '#5686fe',
    '--dsw-static-deepseek-50': '#edf3fe',
    '--dsw-static-deepseek-500': '#4176e6',
    '--dsw-static-deepseek-600': '#4868b2',
    '--dsw-static-deepseek-700-delete': '#2f4c8f',
    '--dsw-static-deepseek-800': '#34415b',
    '--dsw-static-deepseek-900': '#283142',
    '--dsw-static-green-100': '#e6faed',
    '--dsw-static-green-400': '#4ed17e',
    '--dsw-static-green-500': '#22c55e',
    '--dsw-static-green-900': '#233c2c',
    '--dsw-static-neutral-00': '#ffffff',
    '--dsw-static-neutral-100': '#f5f5f5',
    '--dsw-static-neutral-1000': '#000000',
    '--dsw-static-neutral-150': '#ededed',
    '--dsw-static-neutral-200': '#e5e5e5',
    '--dsw-static-neutral-250': '#dcdcdc',
    '--dsw-static-neutral-300': '#d4d4d4',
    '--dsw-static-neutral-400': '#a2a4a6',
    '--dsw-static-neutral-50': '#fafafa',
    '--dsw-static-neutral-500': '#7f8287',
    '--dsw-static-neutral-550': '#65676b',
    '--dsw-static-neutral-600': '#545557',
    '--dsw-static-neutral-700': '#3c3c3d',
    '--dsw-static-neutral-800': '#292929',
    '--dsw-static-neutral-850': '#212123',
    '--dsw-static-neutral-900': '#0f0f0f',
    '--dsw-static-neutral-bluish-00': '#ffffff',
    '--dsw-static-neutral-bluish-100': '#ebeef2',
    '--dsw-static-neutral-bluish-1000': '#0f1115',
    '--dsw-static-neutral-bluish-150': '#e9ecf2',
    '--dsw-static-neutral-bluish-200': '#e1e5ee',
    '--dsw-static-neutral-bluish-300': '#cfd3d6',
    '--dsw-static-neutral-bluish-400': '#adb2b8',
    '--dsw-static-neutral-bluish-50': '#f9fafb',
    '--dsw-static-neutral-bluish-500': '#979da6',
    '--dsw-static-neutral-bluish-60': '#f5f6f7',
    '--dsw-static-neutral-bluish-600': '#81858c',
    '--dsw-static-neutral-bluish-700': '#61666b',
    '--dsw-static-neutral-bluish-75': '#f1f3f5',
    '--dsw-static-neutral-bluish-750': '#43454a',
    '--dsw-static-neutral-bluish-800': '#353638',
    '--dsw-static-neutral-bluish-850': '#2c2c2e',
    '--dsw-static-neutral-bluish-875': '#232324',
    '--dsw-static-neutral-bluish-900': '#1b1b1c',
    '--dsw-static-neutral-bluish-950': '#151517',
    '--dsw-static-red-100': '#fee2e2',
    '--dsw-static-red-400': '#f25a5a',
    '--dsw-static-red-50': '#fef2f2',
    '--dsw-static-red-500': '#ef4444',
    '--dsw-static-red-600': '#ec1313',
    '--dsw-static-red-900': '#570c0c',
  },
  aliases: {
    light: {
// ...
      '--dsw-alias-label-primary-foreground': 'var(--dsw-static-neutral-bluish-00)',
      '--dsw-alias-label-primary-inverted': 'var(--dsw-static-neutral-bluish-00)',
      '--dsw-alias-label-secondary': 'var(--dsw-static-neutral-bluish-700)',
      '--dsw-alias-label-tertiary': 'var(--dsw-static-neutral-bluish-600)',
      '--dsw-alias-markdown-citation': 'var(--dsw-static-neutral-bluish-100)',
      '--dsw-alias-markdown-code-block': 'var(--dsw-static-neutral-bluish-50)',
      '--dsw-alias-markdown-code-block-banner': 'var(--dsw-static-neutral-bluish-50)',
      '--dsw-alias-markdown-code-segment-selected': 'var(--dsw-static-neutral-bluish-00)',
      '--dsw-alias-markdown-code-segment-unselected': 'var(--dsw-static-neutral-bluish-75)',
      '--dsw-alias-markdown-inline-code': 'var(--dsw-static-neutral-bluish-100)',
      '--dsw-alias-markdown-placeholder': 'var(--dsw-static-neutral-bluish-60)',
      '--dsw-alias-markdown-tag': 'var(--dsw-static-neutral-bluish-75)',
      '--dsw-alias-scrollbar-bg-l1': 'var(--dsw-static-neutral-200)',
      '--dsw-alias-scrollbar-bg-l2': 'var(--dsw-static-neutral-200)',
      '--dsw-alias-scrollbar-hover-l1': 'var(--dsw-static-neutral-300)',
      '--dsw-alias-scrollbar-hover-l2': 'var(--dsw-static-neutral-300)',
      '--dsw-alias-state-business-primary': 'var(--dsw-static-deepseek-500)',
      '--dsw-alias-state-business-tertiary': 'var(--dsw-static-deepseek-100)',
      '--dsw-alias-state-error-primary': 'var(--dsw-static-red-600)',
      '--dsw-alias-state-error-secondary': 'var(--dsw-static-red-400)',
      '--dsw-alias-state-success-primary': 'var(--dsw-static-green-500)',
      '--dsw-alias-state-success-secondary': 'var(--dsw-static-green-400)',
      '--dsw-alias-state-success-tertiary': 'var(--dsw-static-green-100)',
      '--dsw-alias-state-warn-label': 'var(--dsw-static-amber-600)',
      '--dsw-alias-state-warn-primary': 'var(--dsw-static-amber-500)',
      '--dsw-alias-state-warn-secondary': 'var(--dsw-static-amber-400)',
      '--dsw-alias-state-warn-tertiary': 'var(--dsw-static-amber-100)',
      '--dsw-alias-toast-bg': 'var(--dsw-static-neutral-bluish-800)',
      '--dsw-alias-tooltip-bg': 'var(--dsw-static-neutral-bluish-850)',
      '--dsw-specific-bubble': 'var(--dsw-static-deepseek-50)',
      '--dsw-specific-bubble-highlight': 'var(--dsw-static-deepseek-200)',
      '--dsw-specific-input-major': 'var(--dsw-static-neutral-bluish-00)',
      '--dsw-specific-login-input': 'var(--dsw-static-neutral-bluish-50)',
      '--dsw-specific-menu': 'var(--dsw-alias-bg-layer-3)',
      '--dsw-specific-selector': 'var(--dsw-static-neutral-bluish-60)',
      '--dsw-specific-sidebar-fill': 'var(--dsw-static-neutral-bluish-50)',
      '--dsw-specific-sidebar-nav-item-active': 'var(--dsw-static-neutral-bluish-100)',
      '--dsw-specific-sidebar-nav-item-active-accent': 'var(--dsw-static-deepseek-100)',
      '--dsw-specific-sidebar-nav-item-hover': 'var(--dsw-static-neutral-bluish-75)',
      '--dsw-specific-tip': 'var(--dsw-static-neutral-bluish-60)',
    },
// ...
  },
}